config_mbr: Fix partition table generation
BZ: 86101

[PORT FROM OTC] - OrigChange-Id: I4a8d173bef231d10e8fbcd0d8aecf54d1cf92474

config_mbr() doesn't produce a valid partition table if there was not
one to begin with; the magic sequence 0xAA55 is not written to the last
2 bytes of the MBR. If this MBR signature is missing, the kernel will
refuse to inspect the partition table.

We add a step to config_mbr which blanks out bytes 440-512 of the MBR
and properly writes the MBR signature before doing anything else.
This fixes a problem where diskinstaller had to be run twice if there
was no partition table present on the target volume.

Change-Id: I5157930b48bf7ba9cae7432beeb86063ae2b84a8
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Signed-off-by: Guilhem IMBERTON <guilhem.imberton@intel.com>
Signed-off-by: Laurent FERT <laurent.fert@intel.com>
Reviewed-on: http://android.intel.com:8080/92002
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Wood, Brian J <brian.j.wood@intel.com>
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>