[PORT FROM MAIN JB-MR2] Use progressive offset in buffer for successive writes
BZ: 88200

The code for copying from one raw file to another looped
back and tried again with a suitably smaller number of bytes
if the entire buffer was not written on the first try.
However, it failed to advance an offset in the buffer past
the bytes that were written.  This could conceivably happen
under rare circumstances, leading to a corrupted disk image.

This fix defines an offset into the buffer, advancing it by
the number of bytes written on each successive retry.

Change-Id: I703926c41089171671dedf35af0a18b507004e3f
Orig-Change-Id: I3632c59d14fd7f6a6f4252acd23ddb56a6816154
Category: aosp improvement
Domain: AOSP-Framework-CoreService
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Paul Drews <paul.drews@intel.com>
Signed-off-by: Jacek Lawrynowicz <jacek.lawrynowicz@intel.com>
Signed-off-by: Laurent FERT <laurent.fert@intel.com>
Signed-off-by: Guilhem IMBERTON <guilhem.imberton@intel.com>
Signed-off-by: Chetan Rao <chetan.rao@intel.com>