[PORT FROM MAIN] Use progressive offset in buffer for successive writes
BZ: 88200

[PORT FROM OTC]
Orig-Change-Id: I76b8e0b3c0d3b0c32bd2114d0867ac7489808279

The code for copying from one raw file to another looped
back and tried again with a suitably smaller number of bytes
if the entire buffer was not written on the first try.
However, it failed to advance an offset in the buffer past
the bytes that were written.  This could conceivably happen
under rare circumstances, leading to a corrupted disk image.

This fix defines an offset into the buffer, advancing it by
the number of bytes written on each successive retry.

Change-Id: I96babbb08d163d54085fae47400ded4193fac624
Orig-Change-Id: I3632c59d14fd7f6a6f4252acd23ddb56a6816154
Signed-off-by: Paul Drews <paul.drews@intel.com>
Signed-off-by: Jacek Lawrynowicz <jacek.lawrynowicz@intel.com>
Signed-off-by: Laurent FERT <laurent.fert@intel.com>
Signed-off-by: Guilhem IMBERTON <guilhem.imberton@intel.com>
Reviewed-on: http://android.intel.com:8080/92612
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Wood, Brian J <brian.j.wood@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>