adb: reserve a wakelock when clients are connected over socket
USB OTG connections already have a wakelock reserved, but if a network
socket connection is made, the device can unexpectedly suspend
in the middle of a session. The lock is released when the client
disconnects.

We have found in testing that most of the USB Ethernet adaptors
we are using do not properly support wake-on-LAN.

Issue: ADEV-21
Change-Id: Ibe7ca1202d45863464d29633381281aa2919c82d
Category: AOSP Improvement
Domain: AOSP-adb
Upstream-candidate: Yes
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>