[PORT FROM MAIN]fastboot: fix fastboot timeout issue of Windows host side
BZ: 48044

Normally fastboot follows the procedure that host sends a command
to device and device sends back response after the command
is executed.

But sometimes device spends too long time to execute the command
so that timeout error occurs before host receives the response.

This patch fixes the issue by aligning with the solution of ADB.

Change-Id: If4302cfe84ce9693bfcc1b033c506be24bfc272b
Signed-off-by: Jiebing Li <jiebing.li@intel.com>
Reviewed-on: http://android.intel.com:8080/76530
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Meng, Zhe <zhe.meng@intel.com>
Tested-by: Meng, Zhe <zhe.meng@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>