init: HACK: Try multiple times to open console
If the console is set to be a USB-to-serial adapter, there is
a bootup race for the creation of its associated device nodes
and init's one-time console detection. 

If the console can't be opened, try polling up to 5 seconds. Log
an error if it ultimately fails.

A better solution would be to reimplement the console functionality
to be hot-pluggable, but this would require much more invasive changes
to make it robust.

Issue: ADEV-21
Change-Id: I347bf75e066fd3663ad63e1e3081e6de75aea1f1
Domain: AOSP-init
Category: AOSP Improvement
Source: Internal
Upstream-candidate: No
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>