[PORT FROM MAIN]ueventd/wifi: add support for module args
BZ: 97273

ueventd modprobes a module when it receives the uevent from the kernel, but
unfortunately there is no way to add specific module arguments at this time.

This adds support for kernel module arguments by specifying it in uevent.*.rc
files. The syntax is:

  mod:%mod_name %mod_args

where %mod_name is to be replaced by the name of the module, and %mod_args
by the arguments.

Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes

Change-Id: I51b7044a31ee730edab837329408f33f93e88d69
Orig-Change-Id: Icebf9f22b2c0f772d106f5e5c9d2bbf49e310e9d
Signed-off-by: Quentin Casasnovas <quentinx.casasnovas@intel.com>
Signed-off-by: Jeremie GARCIA <jeremie.garcia@intel.com>