ueventd: add support for module args
BZ: 97273

ueventd modprobes a module when it receives the uevent from the kernel, but
unfortunately there is no way to add specific module arguments at this time.

This adds support for kernel module arguments by specifying it in uevent.*.rc
files. The syntax is:

  mod:%mod_name %mod_args

where %mod_name is to be replaced by the name of the module, and %mod_args
by the arguments.

Change-Id: Ib8367764503e553aa3b39aa272f8e693b5136c4c
Signed-off-by: Quentin Casasnovas <quentinx.casasnovas@intel.com>
Reviewed-on: http://android.intel.com:8080/96546
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>