init: init process - Ensure console entry to graphics mode
BZ: 91348

After the init process attempts (and fails) to display initlogo.rle,
ensure that the console is set to graphics mode.

If file initlogo.rle was not present (and it is usually not), then
the console would have been left in text mode (as opposed to graphics
mode).  This has ramifications due to console blanking timeout and
the mode setting applied when unblanking.

Change-Id: Ia7238aa9996952e41f322f8755752558f8adc0bd
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>
Reviewed-on: http://android.intel.com:8080/96483
Reviewed-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Tested-by: Gummadi, Latha C <latha.c.gummadi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>