Increase number of available properties
BZ: 80168

We have reached the maximum number of available properties:
PA_COUNT_START. Currently this number is set to 247. This patch
increases this value to 496. The resulting size of the property area
will be 64k (PA_SIZE) instead of 32k.

Note: this values had already been changed on different concurrent
phones.

Change-Id: I093348092b7d8ce01537942c85a14c1a4b3a10e9
Signed-off-by: Clement Calmels <clement.calmels@intel.com>
Reviewed-on: http://android.intel.com:8080/97484
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Shen, JianchunX <jianchunx.shen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>