cleanup ADB's debug output
The server was using stderr to talk back to the client to ACK it
starting correctly, which made the debug output from the server get
lost (before it later gets redirected to a file).  Change it to stdout
and add a fflush() to force the output.

Add the pid to the output so you can tell which process the output is
coming from.

Add basename to the __FILE__ part of the output to remove the unnecessary
and redundant system/core/adb path info from every output line.

Remove the duplicated fdevents debug output macro so it gets redirected
into the server log file like it should.

Category: AOSP Improvement
Domain: AOSP-adb
Origin: internal
Upstream-Candidate: Yes
Change-Id: I8eaf3c8ccbca62f907b0ee7b52a0e179db5ff82d
Signed-off-by: Ken Lierman <ken.lierman@windriver.com>