[PORT FROM MAIN] ANDROID-COS: Add a COS-MOS transition threshold with indication
BZ: 49173

[DO NOT UPSTREAM]

This patch adds an additional check in COS that allows boot to MOS only
with atleast a minimum capacity. As the minimum thresholds for booting to
COS(3.3V) & MOS(3.4V) are different, we should not allow the transition from
COS to MOS before a certain minimum battery capacity.

To improve the user experience, this patch also adds an indication in the
COS-display that it is still not ready to transition from COS to MOS.
That way the user does not have to keep trying.
An yellow-bang is added along with the animation as long as the battery
capacity hasn't reached the minimum required threshold.

Orig-Change-Id: Ia4a534ddb0151fafe079e663e5cea92d9a7ce6c9

Change-Id: Ib68a7c01d839ae42211ed182ec7807cea307f2d1
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>