Allow media process to update AudioComms properties
In order to be able to select a specific XML configuration for the
parameter-framework at runtime, the Android property Audiocomms.PFW.ConfPath
must be updated. However the process where the updated can be achieved
is part of the group media, which is not allowed to update this property.

This patch aims to allow media processes to update Android properties
whose name start with "AudioComms" and "audiocomms".

Issue: 103991
Change-Id: I444e3d510c9c14351b44906e536a3a4eecf540e1
Category: Device Enablement
Domain: AOSP-init
Origin: Internal
Upstream-candidate: No
Signed-off-by: Frederic Boisnard <fredericx.boisnard@intel.com>