[WM8958] Allow media process to update AudioComms properties
BZ: 103991

In order to be able to select a specific XML configuration for the
parameter-framework at runtime, the Android property Audiocomms.PFW.ConfPath
must be updated. However the process where the updated can be achieved
is part of the group media, which is not allowed to update this property.

This patch aims to allow media processes to update Android properties
whose name start with "AudioComms" and "audiocomms".

Change-Id: I6f7b147a63202dd11da744236add9647f7929acf
Orig-Change-Id: I0d660eb1fb7e87f9c85bce1c7870f7ee36ddf71a
Category: device enablement
Domain: AudioComms-WM8958
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Frederic Boisnard <fredericx.boisnard@intel.com>