[PORT FROM MAIN] [WM8958] Allow media process to update AudioComms properties
BZ: 103991

In order to be able to select a specific XML configuration for the
parameter-framework at runtime, the Android property Audiocomms.PFW.ConfPath
must be updated. However the process where the updated can be achieved
is part of the group media, which is not allowed to update this property.

This patch aims to allow media processes to update Android properties
whose name start with "AudioComms" and "audiocomms".

Change-Id: I0d660eb1fb7e87f9c85bce1c7870f7ee36ddf71a
Orig-Change-Id: I96c4e32491a165334a0bd0634f488442a7126ae0
Signed-off-by: Frederic Boisnard <fredericx.boisnard@intel.com>