[PORT FROM MAIN] ANDROID-COS: Fix the screen always on issue when charging in COS
BZ: 55126

We need to power off the screen when charging in COS, and power
on the screen to draw battery animation when input event happen.
It is for efficient charging. Currently power on screen when power
key press and charger removal.

This patch also clears display buffer during display off.
Without this, last drawn image can momentarily flicker
when the screen turns-on again. (BZ: 91446)

Orig-Change-Id: I4f1efcf18ee6a24ed7dace8fca1c12195d8410d9

Change-Id: Ic51d547eb19ed6ad71b6c196e1fc4bb65473a04f
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>