[PORT FROM MAIN]adbd: adb root should terminate adbd only if it's debuggable
BZ: 110652

adb root command will terminate adbd when closing the socket if current
user id is not root. This works for userdebug build, as adb root causes
re-enumeration.

But for user build, adb root command won't cause re-enumeration, and if
adbd is terminated and restarted, it will be in offline state, thus it
won't send any thing to host.

Change-Id: Ie2e844221f3474402a7e2a8535f84ee86a631994
Orig-Change-Id: I68825bc79358d87a3e1b41f4658cd484e8eadb1e
Signed-off-by: jzhuan5 <jin.can.zhuang@intel.com>
Reviewed-on: http://android.intel.com:8080/122566
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Huang, Bo B <bo.b.huang@intel.com>
Tested-by: Wang, Ran <ran.wang@intel.com>
Reviewed-by: Wang, Yu Y <yu.y.wang@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>