[PORT FROM OTC] Fork ueventd to handle firmware in it's child
BZ: 128517

In ueventd we loop for device events and we treat device drivers and
firmware loading requests in the same loop. If a driver makes a request
for firmware, ueventd can't load the firmware because it's blocked in
driver loading.
Therefore, we need to take care of driver loading and firmware loading
in separate processes, so we don't block eachother.

Change-Id: I6b7d6387bb1d7bfd3239658b290ded818f0b75fb
For: AXIA-1977
Issue: AXIA-1977
Signed-off-by: Robert Chiras <robert.chiras@intel.com>
Reviewed-on: http://android.intel.com:8080/123130
Reviewed-by: Falempe, Jocelyn <jocelyn.falempe@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Casasnovas, QuentinX <quentinx.casasnovas@intel.com>
Reviewed-by: Imberton, Guilhem <guilhem.imberton@intel.com>
Tested-by: Falempe, Jocelyn <jocelyn.falempe@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>