[PORT FROM MAIN]ueventd: Overload the do_setprop builtin
BZ: 125765

This patch overloads the do_setprop builtin so we can call it from another
process than init, particularly from ueventd. This extra function makes sure to
keep the properties in a temporary area if the init process hasn't initialized
the property_service at the time the property needs to be defined. This is
required so we can later on add ueventd actions which will set properties on
receival of certain modalias, which can happen really early in the boot
process.

Change-Id: I18a0d5c3d5d9a82087434eeb6f4b95de734290be
Orig-Change-Id: I56f74fb2d4a0f164e329ba2ddff21350c8363ca8
Category: device enablement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Quentin Casasnovas <quentinx.casasnovas@intel.com>
Reviewed-on: http://android.intel.com:8080/118419
Reviewed-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Signed-off-by: Jeremie GARCIA <jeremie.garcia@intel.com>