SDCARD FUSE filesystem "Fast Path" userland optimization
BZ: 128258
This optimization allows the SDCARD FUSE filesystem to take
advantage of the "Fast Path" kernel driver modification.

This FUSE modification allows FUSE filesystems to set the FOPEN_FAST_PATH flag
when opening files. When this flag is set the FUSE kernel module will
handle read/write/flush/sync calls directly for that file, and will not call
the userland FUSE filesystem.

A FUSE filesystem shall only set this flag when no additional
processing/tracking is required for the specified callbacks. For files that
do not have the flag set this option has no effect.

The optimization is performed on the AIO callback, and will not affect
mmap'd files (readpage/writepage callbacks). For supported operations this
provides 2X-5X lower clocks per I/O than the standard FUSE processing.

This is dependant on the kernel FUSE optimization.

Motivation: The Intel Social Camera and Gallery (SDI team) requested this
feature after experiencing performance issues with their FUSE filesystem
(Limitless Storage). After researching the issue the root cause was
determined to be related to overhead of calling to userland. The team
requested the ability to enable/disable the optimization per file leading
to a FUSE open flag implementation. This patch enables the same optimization
for all sdcard access.

Related: http://android.intel.com:8080/#/c/123224 and
http://android.intel.com:8080/#/c/110329/

Change-Id: Ia3aaa36b6cc40af7d3e923f31cc1058bc9006d5f
Signed-off-by: Jason Akers <jason.b.akers@intel.com>
Reviewed-on: http://android.intel.com:8080/123225
Tested-by: Sun, Jianhua <jianhua.sun@intel.com>
Reviewed-by: Gao, Yunpeng <yunpeng.gao@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>