BACKPORT Handle screendumps for all framebuffer sizes
Port from google: https://android-review.googlesource.com/#/c/66261/

readx() treats a partial read as an error but also consumes the data,
Now exactly the amount of data needed for the screendump is requested.

This bug showed up for framebuffers that were not a multiple of 640
bytes.

Also fix a compiler warning related to handing pipe() failure.

Issue: GMIN-264
Change-Id: I8ab8132a5da658f22b14b657968a606bbccd2940
Signed-off-by: hanpengx <pengfeix.han@intel.com>
Reviewed-on: https://android.intel.com/182845
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Zhang, Guobin <guobin.zhang@intel.com>
Reviewed-by: Uyyala, Sridhar <sridhar.uyyala@intel.com>
Tested-by: Uyyala, Sridhar <sridhar.uyyala@intel.com>