[PORT FROM MAIN] ANDROID-COS: Add a COS-MOS transition threshold with indication
BZ: 49173

As the minimum thresholds for booting to COS & MOS are different, transition from
COS to MOS before a certain minimum battery capacity should not be allowed.
To improve the user experience, also need an indication in the
COS-display that it is still not ready to transition from COS to MOS.

This patch adds an additional check in COS that allows boot to MOS only with
atleast a minimum capacity. A yellow-bang is added along with the animation as
long as the battery capacity hasn't reached the minimum required threshold.
That way the user does not have to keep trying.

Category: device enableement
Domain: EM.BatteryManagement-Common
Origin: internal
Upstream-Candidate: no, proprietary

Orig-Change-Id: Ib68a7c01d839ae42211ed182ec7807cea307f2d1
Orig-Change-Id: I95e8d6d8fb909e1d249102f1cf28b7ad4171e008

Change-Id: Icb4f008df720215b0c7beae9e24fa3bcf8282592
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>