[PORT FROM MAIN] libutils: fix corruption in Vector<> when malloc falied
BZ: 139626

1. When alloc or realloc failed in the function SharedBuffer::editResize,
it would return a NULL pointer, then mStorage would update to be 1 by
SharedBuffer::data() if no pointer check here, which is an obviously
wrong address, and would cause corruption when used it e.g. in capacity().
So add the pointer check here for the return value of SharedBuffer::editResize,
if it's NULL do not use it to update mStorage, to avoid the value of mStorage
polluted.
2. when alloc or realloc falied in _grow & _shrink function, mStorage keep
the original value, so mCount should not be updated here.
Otherwise, mStorage might be 0 but mCount>0, so a corruption would happend
when it try to delete items from the Vector since mCount>0.

Orig-Change-Id: I7c3814e843c459834ca5eed392e8d63d1cb7d2d8
Category: aosp improvement
Domain: AOSP-Framework-OS
Origin: internal
Upstream-Candidate: yes
Change-Id: I9debbd4bae8de152cabbe1c7f6ba7b8e8b0abb36
Signed-off-by: Shuo Gao <shuo.gao@intel.com>