Change call to flash function to allow sparse files
BZ: 146204

Using do_flash function will allow flashing with the sparse
option -S.
The sparse option is useful to avoid flashing huge files in
one unique step.
In the case of system.img it will avoid allocating a single
block of more than 1 GB of RAM.

In a first step the option will have to be used with
-S with the size packets to be transfered.

Additional patch will be made in the future that will
automatically sparse or not the files to be flashed function
of the available memory. Done with bug 157254.

Change-Id: Ie1ed5a5a5bc6ced3e18a46b42739501ca5d20887
Category: feature differentiation
Domain: PUPDR-Framework-core
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Michel Jauffres <michelx.jauffres@intel.com>