property_service: add more rights for mediaserver
BZ: 158239

Problem:
Camera HAL needs to let the camera service know whether zsl mode
should be disabled or not. This information is unfortunately
provided via a system property called "camera.disable_zsl_mode".
Because the information needs to be provided per camera device,
depending on whether the device supports RAW captures or not,
a common makefile property will not work since it is common that
the device has multiple cameras of which some are capable of RAW
captures and others are not. Thus the Camera HAL
needs to write the property depending on which camera is opened.

Solution:
Writing camera.* properties to the property mechanism is enabled
for mediaserver.

Change-Id: Ie07fd4da5a3d4475c21f1d027a6a6a797d5b5115
Category: device enablement
Domain: Camera-Framework
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>