fs_mgr: support include statements
When parsing an fstab file, if 'include <filename>' lines are found
they will be parsed as well.

This has been found to be useful when there is a main fstab containing
the internal partition layout (which tends to not change much or at all
between boards) and a secondary file included which has the removable
storage/vold configuration (which changes often based on peripherals).

Upstream-Candidate: Yes
Domain: AOSP-fs_mgr
Category: AOSP Improvement
For: ADEV-476
Change-Id: Id845916b938a189ce81fad21bdf98be869fa57b9
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>