Pass -a command-line option to ADB server.
BZ: 162711

ADB is used to provide a bridge between the GD2 (Gfx Debug) debugger
and a GLink server running on an Android target.

For this to work, ADB must be able to accept TCP/IP connections from
any machine, not just the local host.  In JellyBean, it was necessary
to patch the ADB code to get this to work, but in KitKat there is a
new "-a" command-line option to ADB which allows this.  However, when
ADB spawns a new instance of itself as a server process, it does not
pass on the "-a" option, so the server will not accept connections
from other machines.

This change fixes this problem by passing the "-a" option to the
server process on its command line.

Change-Id: I29895727dc392500a1543f6c470e5568299a6d5c
Category: aosp improvement
Domain: USB-ADB
Origin: internal
Upstream-Candidate: yes
Signed-off-by: rstuckey <richard.stuckey@intel.com>