EM/THERMAL: Handle CPU hotplug 'online' event
BZ: 191851

On a CPU hotplug, the 'online' event is not handled
by the UEvent handler in Android. This patch makes the
UEvent code handle the 'online' event; and provide
'system' permission to the path configured in
ueventd.rc files on receiving this online event.

Change-Id: I3b713040346fd310658676f6bde053af57361284
Category: device enablement
Domain: ThermalManagement-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>
Signed-off-by: Clement Calmels <clement.calmels@intel.com>
Signed-off-by: Robert Jarzmik <robert.jarzmik@intel.com>