dynamic detection/setting of the build fingerprint
It is a Google requirement that the build fingerprint
of an OEM device is unique to that device and is compliant
with the CDD rules for this (section 3.2.2)

This patch adds the code to set these properties at the
earliest possible time (prior to parsing the boot/build properties).
It has to happen this early (and not late in the boot) since these
are ro.* properties, which cannot be changed once set.

Change-Id: I4d038a4b46cbf9caa28caca10f4a6a1489410cf4
Reviewed-on: https://android.intel.com/190597
Reviewed-by: Mahalingam, Ganesh <ganesh.mahalingam@intel.com>
Reviewed-by: Van De Ven, Arjan <arjan.van.de.ven@intel.com>
Tested-by: Mahalingam, Ganesh <ganesh.mahalingam@intel.com>