healthd: fix back trace issues during charger plug-in/plug-out and S3
Issue: BMDA-1157

When a charger in connected/disconnected, the ac charger module posts
an battery event. And healthd module(which is listening to these battery
event) upon receiving the event, calls the get_property of battery module.
The get_property function call take some time to completion since it
involves calls to EC. And since this get_property is sysfs based call, it
can be called anytime.
But in the mean time, due to inactivity in the system, the system suspend
is triggered, which results in back trace.

Hence the fix is to hold a wake_lock, so that suspend/resumes cycles are
prevented till the healthd gets the battery status.

Change-Id: I85c7bcb89bae94754d95d9cd21610eadc2a5cbed
Signed-off-by: Radhesh Fadnis <radheshx.fadnis@intel.com>
Reviewed-on: https://android.intel.com/211873
Reviewed-by: sys_pccgand <sys_pccgand@intel.com>
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Hiremath, Vijay P <vijay.p.hiremath@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>