liblog: limit throughput with LIBLOG_THROUGHPUT_LIMIT
Issue: CAR-1273
Issue: CAR-685

If too many messages are logged during a short period of time
some of them might be lost. This can result in failures for
applications that read and analyze the log messages.

This patch introduces a process specific environment variable

  LIBLOG_THROUGHPUT_LIMIT=<msg_count>/<delay_millis>

If it is set then the process delays for delay_millis
milliseconds after printing every msg_count number of messages.

On an aosp_x86_64 emulator the following setting should be
enough to set if only one process is actively logging.

  LIBLOG_THROUGHPUT_LIMIT=201/200

This means that the process will print at most 201 * 1/0.200 =
= 1005 messages per second.

Note: the environment variable is chosen over an Android system
property, because the property getter function is implemented in
libcutils that depends on liblog. Using system property in liblog
would result in a circular dependency, which is not good.
Moreover, environment variables are process specific and allow
finer tuning particular processes.

Category: Device Enablement
Domain: AOSP-liblog
Origin: internal
Upstream-Candidate: Yes
Change-Id: I59794dc3f5e90365dd038afab9648a4d4e767cd1
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>
Reviewed-on: https://android.intel.com/285748
Reviewed-by: Katkov, Serguei I <serguei.i.katkov@intel.com>
Reviewed-by: Kazantsev, Maxim <maxim.kazantsev@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lupusoru, Razvan A <razvan.a.lupusoru@intel.com>
Reviewed-by: lab_aqa <lab_aqa@intel.com>
Reviewed-by: Beyler, Jean Christophe <jean.christophe.beyler@intel.com>
Reviewed-by: Zolotukhin, Konstantin A <konstantin.a.zolotukhin@intel.com>
Reviewed-by: Yakovlev, Andrey <andrey.yakovlev@intel.com>
Reviewed-by: Mendell, Mark P <mark.p.mendell@intel.com>
Reviewed-by: Chen, Dong-Yuan <dong-yuan.chen@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>
Reviewed-on: https://android.intel.com/324675
Tested-by: cactus <cactus@intel.com>
Reviewed-by: Semukhina, Elena V <elena.v.semukhina@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-27381
Reviewed-on: https://android.intel.com:443/346238
Reviewed-on: https://android.intel.com:443/354405