Allow DMI "board_name1" to populate ro.product.model
IRDA is working with IBV's to allow them to set in the BIOS an OEM specific
human-pretty marketing name in the DMI variable "board_name1".  We expose
that in ro.product.model instead of the device name.  The device name is
the engineering name used in the very short fingerprint string and is
"ugly".  Allowing a "pretty" ro.product.model means a human looking at
Settings->About Tablet->Model number will see a pretty string.

On Nexus7 devices ro.product.model is set to "Nexus 7", where
ro.product.device is set to "flo".

Issue: IRDA-2688
Change-Id: If026a0a7a35f30251de2340c31b9a113e7cde772
Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>
Reviewed-on: https://android.intel.com:443/307062