Don't modify ro.board.platform
This value is not being used for anything IRDA related.
The 3 main values of ro.product.brand, ro.product.device and
ro.product.name are being set and fingerprint is also explicitly
modified. Removing this override helps in setting the same
board.platform for all variants of COHO (dev-kernel, IES, etc)

This fixes init finding the right libhwcomposer.so file

Change-Id: I087fd24de5b03d8ff61dcc3f898cd0fd0cce9769
Signed-off-by: Nisha Kumar-Mayernik <nisha.kumar-mayernik@intel.com>
Reviewed-on: https://android.intel.com/353094